Move TestedCasesDetailsCard to components/index/cards
